Spray Paint Plus v1.11.0: paired client and server settings
Every capability now has a client half and a server half and works only when both
allow it. 35 settings across eleven groups, up from 18 across five.

Color cycling becomes a three-mode dropdown (cannot change / within paint family /
all colors), the stricter of the two halves winning. Color picking becomes its own
paired setting. Enable Metallic Paints is removed; the family mode replaces it.

Adds join-time config sync so a client learns the server's halves, a consolidated
console message naming blocked functions at join, and a warning the first three
times a blocked function is actually used. The per-player preference mask rides
the widened PaintModifierMessage; SettingBlockedNotice carries the server-detected
case back to one client.

Fixes two defects the settings audit surfaced: Extra Paintable Structures and
Glow Paint were already effectively client-and-server but mismatches were silent,
and a glowing object recoloured with a plain can stopped glowing for any client
with glow disabled while everyone else still saw it.

All player settings reset to defaults on this update, as at v1.5.0, because
BepInEx identifies an entry by its section plus key and nearly every entry moved.

Verified headlessly on the dedicated server: 111 assertions covering the merge
truth table, the cycling ladder in all nine combinations, authority resolution in
all three network roles, the full 16x16 paint-family matrix, and the effective
settings log line. Scenario spp-settings-merge-verify; spp-dlc-gate-verify
retargeted onto the new signatures.

+++ Mods/SprayPaintPlus/SprayPaintPlus/PaintModifierMessage.cs
@@ -4,31 +4,41 @@ using LaunchPadBooster.Networking;
 namespace SprayPaintPlus
 {
     /// <summary>
-    /// Client -> Server message: player's modifier key state changed while holding a spray can.
-    /// Bit 0 = Shift (single item paint, invert already applied), Bit 1 = Ctrl (checkered).
+    /// Client -> Server message: the sender's client-half preference mask changed while
+    /// holding a spray can or spray gun.
+    ///
+    /// Modifiers carries the sender's WHOLE client half, not only the two live modifier
+    /// keys it carried before v1.11.0: every client-side setting the server has to merge
+    /// before it acts on that player's paint (the eleven network-painting toggles, glow
+    /// paint, unlimited uses) rides along with them. Bit positions live in
+    /// SettingsMerge.PlayerPrefs, which is the single source of truth for the layout.
+    /// Appending a bit there is safe; renumbering one is not, because both ends of this
+    /// message must agree and MOD.Networking.Required only guarantees a matching mod
+    /// version, not a matching interpretation of an old one.
+    ///
     /// PlayerHumanId is the sender's own controlled Human ReferenceId; the server keys
     /// PlayerModifiers by that id because vanilla paint messages identify the actor
     /// by AttackParentId (a Human ReferenceId), not by the LaunchPadBooster connection id.
     /// </summary>
     public class PaintModifierMessage : INetworkMessage
     {
-        public byte Modifiers;
+        public ushort Modifiers;
         public long PlayerHumanId;
 
         public void Serialize(RocketBinaryWriter writer)
         {
-            writer.WriteByte(Modifiers);
+            writer.WriteUInt16(Modifiers);
             writer.WriteInt64(PlayerHumanId);
         }
 
         public void Deserialize(RocketBinaryReader reader)
         {
-            Modifiers = reader.ReadByte();
+            Modifiers = reader.ReadUInt16();
             PlayerHumanId = reader.ReadInt64();
         }
 
         /// <summary>
-        /// Runs on the server when a client's modifier state changes.
+        /// Runs on the server when a client's preference mask changes.
         /// </summary>
         public void Process(long hostId)
         {
